bots/trv-bot/requirements.txt
python-socketio[client]>=5.12,<5.13
numpy